[MOV] mail: move webRecord and saveRecord props (portal to web)
These are the last props in chatter which we don't use in portal.
With this move, there is no usage of chatter attachment uploader in portal
bundle anymore (moved to web folder).

followup of odoo/odoo#159858
