+++ addons/mail/static/src/chatter/web/chatter_patch.js
@@ -6,6 +6,7 @@ import { FollowerList } from "@mail/core/web/follower_list";
 import { useHover } from "@mail/utils/common/hooks";
 import { useDropzone } from "@mail/core/common/dropzone_hook";
 import { isDragSourceExternalFile } from "@mail/utils/common/misc";
+import { useAttachmentUploader } from "@mail/core/common/attachment_uploader_hook";
 
 import { markup, useEffect } from "@odoo/owl";
 import { browser } from "@web/core/browser/browser";
@@ -33,7 +34,9 @@ Chatter.props.push(
     "hasParentReloadOnMessagePosted?",
     "isAttachmentBoxVisibleInitially?",
     "isChatterAside?",
-    "isInFormSheetBg?"
+    "isInFormSheetBg?",
+    "saveRecord?",
+    "webRecord?"
 );
 
 Object.assign(Chatter.defaultProps, {
@@ -61,6 +64,9 @@ patch(Chatter.prototype, {
             showActivities: true,
             showAttachmentLoading: false,
         });
+        this.attachmentUploader = useAttachmentUploader(
+            this.store.Thread.insert({ model: this.props.threadModel, id: this.props.threadId })
+        );
         this.unfollowHover = useHover("unfollow");
         this.followerListDropdown = useDropdownState();
         /** @type {number|null} */
@@ -190,6 +196,18 @@ patch(Chatter.prototype, {
         return _t("Unfollow");
     },
 
+    changeThread(threadModel, threadId) {
+        super.changeThread(...arguments);
+        this.attachmentUploader.thread = this.state.thread;
+        if (threadId === false) {
+            this.state.composerType = false;
+        } else {
+            this.onThreadCreated?.(this.state.thread);
+            this.onThreadCreated = null;
+            this.closeSearch();
+        }
+    },
+
     async _follow(thread) {
         await this.orm.call(thread.model, "message_subscribe", [[thread.id]], {
             partner_ids: [this.store.self.id],
@@ -216,6 +234,16 @@ patch(Chatter.prototype, {
         if (this.state.isAttachmentBoxOpened) {
             this.rootRef.el.scrollTop = 0;
             this.state.thread.scrollTop = 0;
+        }
+    },
+
+    async onClickAttachFile(ev) {
+        if (this.state.thread.id) {
+            return;
+        }
+        const saved = await this.props.saveRecord?.();
+        if (!saved) {
+            return false;
         }
     },
 
@@ -312,7 +340,7 @@ patch(Chatter.prototype, {
     },
 
     async unlinkAttachment(attachment) {
-        await super.unlinkAttachment(attachment);
+        await this.attachmentUploader.unlink(attachment);
         if (this.props.hasParentReloadOnAttachmentsChanged) {
             this.reloadParentView();
         }
