+++ addons/mail/static/src/chatter/web_portal/chatter.js
@@ -1,5 +1,4 @@
 import { AttachmentList } from "@mail/core/common/attachment_list";
-import { useAttachmentUploader } from "@mail/core/common/attachment_uploader_hook";
 import { Composer } from "@mail/core/common/composer";
 import { Thread } from "@mail/core/common/thread";
 import { useMessageHighlight } from "@mail/utils/common/hooks";
@@ -34,7 +33,7 @@ export class Chatter extends Component {
         FileUploader,
         SearchMessagesPanel,
     };
-    static props = ["threadId?", "threadModel", "webRecord?", "saveRecord?"];
+    static props = ["threadId?", "threadModel"];
     static defaultProps = { threadId: false };
 
     setup() {
@@ -50,9 +49,6 @@ export class Chatter extends Component {
             thread: undefined,
             isSearchOpen: false,
         });
-        this.attachmentUploader = useAttachmentUploader(
-            this.store.Thread.insert({ model: this.props.threadModel, id: this.props.threadId })
-        );
         this.rootRef = useRef("root");
         this.onScrollDebounced = useThrottleForAnimation(this.onScroll);
         this.messageHighlight = useMessageHighlight();
@@ -62,7 +58,7 @@ export class Chatter extends Component {
         });
 
         onMounted(() => {
-            this.changeThread(this.props.threadModel, this.props.threadId, this.props.webRecord);
+            this.changeThread(this.props.threadModel, this.props.threadId);
             if (!this.env.chatter || this.env.chatter?.fetchData) {
                 if (this.env.chatter) {
                     this.env.chatter.fetchData = false;
@@ -75,7 +71,7 @@ export class Chatter extends Component {
                 this.props.threadId !== nextProps.threadId ||
                 this.props.threadModel !== nextProps.threadModel
             ) {
-                this.changeThread(nextProps.threadModel, nextProps.threadId, nextProps.webRecord);
+                this.changeThread(nextProps.threadModel, nextProps.threadId);
             }
             if (!this.env.chatter || this.env.chatter?.fetchData) {
                 if (this.env.chatter) {
@@ -94,10 +90,8 @@ export class Chatter extends Component {
         return [];
     }
 
-    changeThread(threadModel, threadId, webRecord) {
+    changeThread(threadModel, threadId) {
         this.state.thread = this.store.Thread.insert({ model: threadModel, id: threadId });
-        this.state.thread.name = webRecord?.data?.display_name || undefined;
-        this.attachmentUploader.thread = this.state.thread;
         if (threadId === false) {
             if (this.state.thread.messages.length === 0) {
                 this.state.thread.messages.push({
@@ -110,11 +104,6 @@ export class Chatter extends Component {
                     model: threadModel,
                 });
             }
-            this.state.composerType = false;
-        } else {
-            this.onThreadCreated?.(this.state.thread);
-            this.onThreadCreated = null;
-            this.closeSearch();
         }
     }
 
@@ -136,10 +125,6 @@ export class Chatter extends Component {
         this.load(this.state.thread, this.afterPostRequestList);
     }
 
-    async unlinkAttachment(attachment) {
-        await this.attachmentUploader.unlink(attachment);
-    }
-
     onClickSearch() {
         this.state.composerType = false;
         this.state.isSearchOpen = !this.state.isSearchOpen;
@@ -147,16 +132,6 @@ export class Chatter extends Component {
 
     closeSearch() {
         this.state.isSearchOpen = false;
-    }
-
-    async onClickAttachFile(ev) {
-        if (this.state.thread.id) {
-            return;
-        }
-        const saved = await this.props.saveRecord?.();
-        if (!saved) {
-            return false;
-        }
     }
 
     onScroll() {
