+++ addons/mail/static/src/core/web/message_patch.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <templates xml:space="preserve">
     <t t-inherit="mail.Message" t-inherit-mode="extension">
-        <xpath expr="//div[hasclass('o-mail-Message-textContent')]/*[1]" position="before">
+        <xpath expr="//t[@name='bodyAsNotification']" position="replace">
             <t t-if="message.message_type === 'notification' or message.is_transient or message.trackingValues.length > 0" name="hasTrackingValue">
                 <div>
                     <t t-if="message.subtype_description">
@@ -21,11 +21,10 @@
                             </t>
                         </ul>
                     </t>
-                    <div t-if="message.body" class="o-mail-Message-body text-break mb-0 w-100" t-ref="body">
-                        <t t-out="messageSearch?.highlight(message.body) ?? message.body"/>
-                    </div>
+                    <div t-if="message.body" t-call="mail.Message.bodyAsNotification" t-ref="body"/>
                 </div>
             </t>
+            <t t-else="">$0</t>
         </xpath>
         <xpath expr="//div[hasclass('o-mail-Message-avatarContainer')]" position="attributes">
             <attribute name="t-att-class">{ 'cursor-pointer': hasAuthorClickable() }</attribute>
